# Retrieve span outside of loop to improve performance of glyph texture creation